run it

+++ idlib/Heap.cpp
@@ -1362,7 +1362,7 @@ void *Mem_AllocAligned( const size_t size, const align_t align )
     }
     void *mem = mem_heap->AllocateAligned( size, align );
     // make sure the memory is aligned
-    assert( align == ALIGN_NONE || ( ((int)mem) & (align-1)) == 0 );
+    assert( align == ALIGN_NONE || ( ((intptr_t)mem) & (align-1)) == 0 );
     return mem;
 }
 
@@ -2624,7 +2624,7 @@ void *Mem_AllocAligned( const size_t size, const align_t align, const char *file
     }
     void *mem = Mem_AllocDebugMemory( size, fileName, lineNumber, align );
     // make sure the memory is aligned
-    assert( align == ALIGN_NONE || ( ((int)mem) & (align-1)) == 0 );
+    assert( align == ALIGN_NONE || ( ((intptr_t)mem) & (align-1)) == 0 );
     return mem;
 }
 
@@ -2654,7 +2654,7 @@ void *Mem_Alloc16(const int size, const char *fileName, const int lineNumber)
 
 	void *mem = Mem_AllocDebugMemory(size, fileName, lineNumber, true);
 	// make sure the memory is 16 byte aligned
-	assert((((int)mem) & 15) == 0);
+	assert((((intptr_t)mem) & 15) == 0);
 	return mem;
 }
 
@@ -2670,7 +2670,7 @@ void Mem_Free16(void *ptr, const char *fileName, const int lineNumber)
 	}
 
 	// make sure the memory is 16 byte aligned
-	assert((((int)ptr) & 15) == 0);
+	assert((((intptr_t)ptr) & 15) == 0);
 	Mem_FreeDebugMemory(ptr, fileName, lineNumber, true);
 }
